Extracted `trait IntoOperand` and `impl`s into `crate::into_operand`.
